State: Move canCurrentUser to state/selectors
