+++ client/post-editor/editor-term-selector/index.jsx
@@ -10,11 +10,11 @@ import { cloneDeep, findIndex, map, toArray } from 'lodash';
  */
 import TermTreeSelector from 'blocks/term-tree-selector';
 import AddTerm from './add-term';
+import { canCurrentUser } from 'state/selectors';
 import { editPost, addTermForPost } from 'state/posts/actions';
 import { getSelectedSiteId } from 'state/ui/selectors';
 import { getEditorPostId } from 'state/ui/editor/selectors';
 import { getEditedPostValue } from 'state/posts/selectors';
-import { canCurrentUser } from 'state/current-user/selectors';
 
 class EditorTermSelector extends Component {
 	static propTypes = {
