+++ client/state/current-user/test/selectors.js
@@ -12,7 +12,6 @@ import {
 	getCurrentUserLocale,
 	getCurrentUserDate,
 	isValidCapability,
-	canCurrentUser,
 	getCurrentUserCurrencyCode
 } from '../selectors';
 
@@ -182,46 +181,6 @@ describe( 'selectors', () => {
 			}, 2916284, 'manage_foo' );
 
 			expect( isValid ).to.be.false;
-		} );
-	} );
-
-	describe( 'canCurrentUser', () => {
-		it( 'should return null if the site is not known', () => {
-			const isCapable = canCurrentUser( {
-				currentUser: {
-					capabilities: {}
-				}
-			}, 2916284, 'manage_options' );
-
-			expect( isCapable ).to.be.null;
-		} );
-
-		it( 'should return the value for the specified capability', () => {
-			const isCapable = canCurrentUser( {
-				currentUser: {
-					capabilities: {
-						2916284: {
-							manage_options: false
-						}
-					}
-				}
-			}, 2916284, 'manage_options' );
-
-			expect( isCapable ).to.be.false;
-		} );
-
-		it( 'should return null if the capability is invalid', () => {
-			const isCapable = canCurrentUser( {
-				currentUser: {
-					capabilities: {
-						2916284: {
-							manage_options: false
-						}
-					}
-				}
-			}, 2916284, 'manage_foo' );
-
-			expect( isCapable ).to.be.null;
 		} );
 	} );
