+++ client/state/sharing/services/selectors.js
@@ -6,7 +6,7 @@ import { filter } from 'lodash';
 /**
  * Internal dependencies
  */
-import { canCurrentUser } from 'state/current-user/selectors';
+import { canCurrentUser } from 'state/selectors';
 import { isJetpackSite, isJetpackModuleActive } from 'state/sites/selectors';
 
 /**
